--- a/composer.json
+++ b/composer.json
@@ -28,6 +28,7 @@
     "ext-dom": "*",
     "chi-teck/drupal-code-generator": "^1.21.0",
     "composer/semver": "^1.4",
+    "composer/xdebug-handler": "dev-master",
     "consolidation/annotated-command": "^2.8.1",
     "consolidation/config": "^1.0.9",
     "consolidation/output-formatters": "^3.1.12",