[composer/xdebug-handler](https://github.com/composer/xdebug-handler) is a relatively new package. It has no stable release yet. The package provides the xdebug flag composer (COMPOSER_ALLOW_XEBUG) in a general fashion. It works seamlessly with symfony console.

This PR introduces the same feature for drush with DRUSH_ALLOW_XEBUG.

Some non-scientific benchmarks on macOS with PHP 7.1

 ## Before (current default)

```sh
$ export DRUSH_ALLOW_XDEBUG=1 && time ~/Projects/drush9/drush cr
 [success] Cache rebuild complete.
~/Projects/drush9/drush cr  3,72s user 0,40s system 86% cpu 4,773 total
```

**4,773s**


## After (new default)

```sh
$ export DRUSH_ALLOW_XDEBUG=0 && time ~/Projects/drush9/drush cr
 [success] Cache rebuild complete.
~/Projects/drush9/drush cr  1,12s user 0,35s system 70% cpu 2,091 total
```

**2,091s**

---

This is a big BC break. I am not sure if this improves the DX in general, because i often want to use xdebug in conjunction with php-eval or while debugging update hooks. 

Lets discuss :)