@webflo 2017-12-24
BC break because it functions different than before. I think people use drush often do debug something in Drupal? But of course this it not and API change. I think some new docs about drush and xdebug should be sufficient.