@webflo 2017-12-24
@weitzman This is the PR for the launcher. https://github.com/drush-ops/drush-launcher/pull/41