@weitzman 2017-12-28
Just tested this - got similarly impressive results. I think we need a stable release before we add this. Or else Composer users have minimum-stability hassles. here are some docs:


> Drush, like Composer automatically disables Xdebug by default. This improves performance substantially. You may override this feature by setting an environment variable - `DRUSH_ALLOW_XDEBUG=1 drush [command]`. See https://github.com/drush-ops/drush/pull/3272.