@webflo 2018-03-08
New release for drush-launcher: https://github.com/drush-ops/drush-launcher/releases/tag/0.6.0