@weitzman 2018-03-08
Unlike the Launcher, I think Drush can pass its Logger using https://github.com/composer/xdebug-handler/blob/master/README.md#output. The PreflightLog is not a real Logger so I suspect we have to do the xdebug check after this line https://github.com/drush-ops/drush/blob/a7814d2afcf13236455f509679c74f77fc933700/src/Runtime/Runtime.php#L78. 

Also worth setting DRUSH_ as the env variable prefix. Launcher could do that as well. Scroll up on the README to learn more.